+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Fixed
 
-- Fix up `GitCommitRef` `change_counts` field type
-- Add back some required `build` structure Option wrappers
+- Fixed up `GitCommitRef` `change_counts` field type
+- Added back some required `build` structure Option wrappers
+- Fixed formatting of date-time URL parameters
 
 ## [0.5.2]