+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.3]
+
 ## Fixed
 
 - Fixed up `GitCommitRef` `change_counts` field type
@@ -126,7 +128,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/microsoft/azure-devops-rust-api/compare/0.5.2...HEAD
+[Unreleased]: https://github.com/microsoft/azure-devops-rust-api/compare/0.5.3...HEAD
+[0.5.3]: https://github.com/microsoft/azure-devops-rust-api/compare/0.5.2...0.5.3
 [0.5.2]: https://github.com/microsoft/azure-devops-rust-api/compare/0.5.1...0.5.2
 [0.5.1]: https://github.com/microsoft/azure-devops-rust-api/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/microsoft/azure-devops-rust-api/compare/0.4.0...0.5.0